requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.6.0
Flask==3.0.3
gvm-tools==20.10.1
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.4
MarkupSafe==2.1.5
paramiko==2.12.0
pycparser==2.22
PyNaCl==1.5.0
python-gvm==20.12.1
six==1.16.0
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
